requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.3.1
alpaca-py==0.8.1
alpaca-trade-api==3.0.0
async-timeout==4.0.2
attrs==23.1.0
ccxt==3.0.70
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
deprecation==2.1.0
frozenlist==1.3.3
idna==3.4
msgpack==1.0.3
multidict==6.0.4
numpy==1.24.2
packaging==23.1
pandas==1.5.3
pycares==4.3.0
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sseclient-py==1.7.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
